backend/image_generator_service/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt[crypto]>=5.2,<5.4 # Per validazione
django-cors-headers>=3.13,<4.4
gunicorn>=20.1,<22.0
python-dotenv>=0.20,<1.1 # Utile per .env anche se non per le chiavi API
requests>=2.28,<2.32 # Per chiamate API esterne
Pillow>=9.0,<10.3 # Per gestione immagini (anche solo validazione o salvataggio base64)
# Non serve psycopg2-binary se non usiamo DB Postgres qui